server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.7.2
speechrecognition==3.10.0
gtts==2.4.0
pydub==0.25.1
websockets==12.0
python-multipart==0.0.6